Remove unnecessary recipe
